Refactor import paths for trace model and span kind enum
