+++ lib/simple_tracer.dart
@@ -2,5 +2,5 @@
 library;
 
 export 'src/simple_tracer.dart';
-export 'src/span_kind_enum.dart';
-export 'src/trace_model.dart';
+export 'src/model/span_kind_enum.dart';
+export 'src/model/trace_model.dart';
