+++ lib/src/model/span_kind_enum.dart
